Split out Bridge, Host, Node implementations.
It makes it easier to follow if they're in separate files.  They're
actually independent now.
