+++ clusterDeployer.py
@@ -5,7 +5,6 @@ import json
 import xml.etree.ElementTree as et
 import shutil
 from concurrent.futures import ThreadPoolExecutor
-from concurrent.futures import Future
 from typing import Optional
 from typing import Generator
 from typing import Dict
@@ -15,672 +14,16 @@ from typing import Callable
 from typing import Set
 import re
 import logging
-import paramiko
 from assistedInstaller import AssistedClientAutomation
 import host
-from clustersConfig import ClustersConfig, NodeConfig, HostConfig, ExtraConfigArgs
+from clustersConfig import ClustersConfig, ExtraConfigArgs
 from k8sClient import K8sClient
-from nfs import NFS
-import coreosBuilder
-from typing import Tuple
 import common
 from python_hosts import Hosts, HostsEntry
 from logger import logger
 import microshift
 from extraConfigRunner import ExtraConfigRunner
-from host import BMC
-import abc
-
-
-class VirBridge:
-    """
-    Wrapper on top of the libvirt virtual bridge.
-
-    It can be running locally or remote.
-    """
-
-    hostconn: host.Host
-
-    def __init__(self, h: host.Host):
-        self.hostconn = h
-
-    def setup_dhcp_entry(self, cfg: NodeConfig) -> None:
-        if cfg.ip is None:
-            logger.error_and_exit(f"Missing IP for node {cfg.name}")
-        ip = cfg.ip
-        mac = cfg.mac
-        name = cfg.name
-        # If adding a worker node fails, one might want to retry w/o tearing down
-        # the whole cluster. In that case, the DHCP entry might already be present,
-        # with wrong mac -> remove it
-
-        cmd = "virsh net-dumpxml default"
-        ret = self.hostconn.run_or_die(cmd)
-        if f"'{name}'" in ret.out:
-            logger.info(f"{name} already configured as static DHCP entry - removing before adding back with proper configuration")
-            host_xml = f"<host name='{name}'/>"
-            cmd = f"virsh net-update default delete ip-dhcp-host \"{host_xml}\" --live --config"
-            self.hostconn.run_or_die(cmd)
-
-        cmd = "virsh net-dhcp-leases default"
-        ret = self.hostconn.run(cmd)
-        # Look for "{name} " in the output. The space is intended to differentiate between "bm-worker-2 " and e.g. "bm-worker-20"
-        if f"{name} " in ret.out:
-            logger.error(f"Error: {name} found in dhcp leases")
-            logger.error("To fix this, run")
-            logger.error("\tvirsh net-destroy default")
-            logger.error("\tRemove wrong entries from /var/lib/libvirt/dnsmasq/virbr0.status")
-            logger.error("\tvirsh net-start default")
-            logger.error("\tsystemctl restart libvirt")
-            sys.exit(-1)
-
-        host_xml = f"<host mac='{mac}' name='{name}' ip='{ip}'/>"
-        logger.info(f"Creating static DHCP entry for VM {name}, ip {ip} mac {mac}")
-        cmd = f"virsh net-update default add ip-dhcp-host \"{host_xml}\" --live --config"
-        self.hostconn.run_or_die(cmd)
-
-    def _ensure_started(self, api_network: str, bridge_xml: str) -> None:
-        cmd = "virsh net-destroy default"
-        self.hostconn.run(cmd)  # ignore return code - it might fail if net was not started
-
-        cmd = "virsh net-undefine default"
-        ret = self.hostconn.run(cmd)
-        if ret.returncode != 0 and "Network not found" not in ret.err:
-            logger.error_and_exit(str(ret))
-
-        # Fix cases where virsh net-start fails with error "... interface virbr0: File exists"
-        cmd = "ip link delete virbr0"
-        self.hostconn.run(cmd)  # ignore return code - it might fail if virbr did not exist
-
-        cmd = f"virsh net-define {bridge_xml}"
-        self.hostconn.run_or_die(cmd)
-
-        # set interface down before starting bridge as otherwise bridge start might fail if interface
-        # already got an IP address in same network as bridge
-        self.hostconn.run(f"ip link set {api_network} down")
-
-        cmd = "virsh net-start default"
-        self.hostconn.run_or_die(cmd)
-
-        self.hostconn.run(f"ip link set {api_network} up")
-
-    def limit_dhcp_range(self, old_range: str, new_range: str) -> None:
-        # restrict dynamic dhcp range: we use static dhcp ip addresses; however, those addresses might have been used
-        # through the dynamic dhcp by any systems such as systems ready to be installed.
-        cmd = "virsh net-dumpxml default"
-        ret = self.hostconn.run(cmd)
-        if f"range start='{old_range}'" in ret.out:
-            host_xml = f"<range start='{old_range}' end='192.168.122.254'/>"
-            cmd = f"virsh net-update default delete ip-dhcp-range \"{host_xml}\" --live --config"
-            r = self.hostconn.run(cmd)
-            logger.debug(r.err if r.err else r.out)
-
-            host_xml = f"<range start='{new_range}' end='192.168.122.254'/>"
-            cmd = f"virsh net-update default add ip-dhcp-range \"{host_xml}\" --live --config"
-            r = self.hostconn.run(cmd)
-            logger.debug(r.err if r.err else r.out)
-
-    def _network_xml(self, ip: str, dhcp_range: Optional[Tuple[str, str]] = None) -> str:
-        if dhcp_range is None:
-            dhcp_part = ""
-        else:
-            dhcp_part = f"""<dhcp>
-                <range start='{dhcp_range[0]}' end='{dhcp_range[1]}'/>
-                </dhcp>"""
-
-        return f"""
-                <network>
-                <name>default</name>
-                <forward mode='nat'/>
-                <bridge name='virbr0' stp='off' delay='0'/>
-                <ip address='{ip}' netmask='255.255.0.0'>
-                    {dhcp_part}
-                </ip>
-                </network>"""
-
-    def _restart(self) -> None:
-        self.hostconn.run_or_die("systemctl restart libvirtd")
-
-    def _ensure_run_as_root(self) -> None:
-        qemu_conf = self.hostconn.read_file("/etc/libvirt/qemu.conf")
-        if re.search('\nuser = "root"', qemu_conf) and re.search('\nuser = "root"', qemu_conf):
-            return
-        self.hostconn.run("sed -e 's/#\\(user\\|group\\) = \".*\"$/\\1 = \"root\"/' -i /etc/libvirt/qemu.conf")
-        self._restart()
-
-    def configure(self, api_network: str) -> None:
-        hostname = self.hostconn.hostname()
-        cmd = "systemctl enable libvirtd --now"
-        self.hostconn.run_or_die(cmd)
-
-        self._ensure_run_as_root()
-
-        # stp must be disabled or it might conflict with default configuration of some physical switches
-        # 'bridge' section of network 'default' can't be updated => destroy and recreate
-        # check that default exists and contains stp=off
-        cmd = "virsh net-dumpxml default"
-        ret = self.hostconn.run(cmd)
-
-        if "stp='off'" not in ret.out or "range start='192.168.122.2'" in ret.out:
-            logger.info("Destoying and recreating bridge")
-            logger.info(f"creating default-net.xml on {hostname}")
-            if hostname == "localhost":
-                contents = self._network_xml('192.168.122.1', ('192.168.122.129', '192.168.122.254'))
-            else:
-                contents = self._network_xml('192.168.123.250')
-
-            bridge_xml = os.path.join("/tmp", 'vir_bridge.xml')
-            self.hostconn.write(bridge_xml, contents)
-            # Not sure why/whether this is needed. But we saw failures w/o it.
-            # Without this, net-undefine within ensure_bridge_is_started fails as libvirtd fails to restart
-            # We need to investigate how to remove the sleep to speed up
-            time.sleep(5)
-            self._ensure_started(api_network, bridge_xml)
-
-            self.limit_dhcp_range("192.168.122.2", "192.168.122.129")
-
-            self._restart()
-
-            # Not sure why/whether this is needed. But we saw failures w/o it.
-            # We need to investigate how to remove the sleep to speed up
-            time.sleep(5)
-
-
-class ClusterNode:
-    """
-    Base class for all k8s-like nodes to inherit from.
-
-    At a minimum, a correct implementation should provide:
-    - a start() method: to provision the node.
-    - a has_booted() method: to determine if the node has booted
-    - a post_boot() method: to execute actions after the node has been
-      provisioned and booted for the first time
-    """
-
-    config: NodeConfig
-    future: Future[Optional[host.Result]]
-    dynamic_ip: Optional[str] = None
-
-    def __init__(self, config: NodeConfig):
-        def empty() -> Future[Optional[host.Result]]:
-            f: Future[Optional[host.Result]] = Future()
-            f.set_result(None)
-            return f
-
-        self.config = config
-        self.future = empty()
-
-    def ip(self) -> str:
-        if self.config.ip is not None:
-            return self.config.ip
-        if self.dynamic_ip is None:
-            logger.error_and_exit(f"Node {self.config.name} has no IP address")
-        return self.dynamic_ip
-
-    @abc.abstractmethod
-    def start(self, iso_or_image_path: str, executor: ThreadPoolExecutor) -> None:
-        pass
-
-    @abc.abstractmethod
-    def has_booted(self) -> bool:
-        pass
-
-    @abc.abstractmethod
-    def post_boot(self, desired_ip_range: Tuple[str, str]) -> bool:
-        pass
-
-    def teardown(self) -> None:
-        pass
-
-    def set_password(self, user: str = "root", password: str = "redhat") -> None:
-        rh = host.RemoteHost(self.ip())
-        rh.ssh_connect("core")
-        rh.run_or_die(f"echo {user}:{password} | sudo chpasswd")
-
-    def print_logs(self) -> None:
-        rh = host.RemoteHost(self.ip())
-        logger.info(f"Gathering logs from {self.config.name}")
-        logger.info(rh.run("sudo journalctl TAG=agent --no-pager").out)
-
-    def _verify_package_is_installed(self, package: str) -> bool:
-        rh = host.RemoteHost(self.ip())
-        rh.ssh_connect("core")
-        ret = rh.run(f"rpm -qa | grep {package}")
-        return not ret.returncode
-
-    def health_check(self) -> None:
-        required_packages = ["kernel-modules-extra"]
-        missing_packages = [p for p in required_packages if not self._verify_package_is_installed(p)]
-        for p in missing_packages:
-            logger.error(f"Required rpm '{p}' is not installed")
-        if any(missing_packages):
-            sys.exit(-1)
-
-
-class VmClusterNode(ClusterNode):
-    hostconn: host.Host
-    install_wait: bool = True
-
-    def __init__(self, h: host.Host, config: NodeConfig):
-        super().__init__(config)
-        self.hostconn = h
-
-    def ip(self) -> str:
-        assert self.config.ip is not None
-        return self.config.ip
-
-    def setup_vm(self, iso_or_image_path: str) -> host.Result:
-        disk_size_gb = self.config.disk_size
-        if iso_or_image_path.endswith(".iso"):
-            options = "-o preallocation="
-            if self.config.is_preallocated():
-                options += "full"
-            else:
-                options += "off"
-
-            os.makedirs(os.path.dirname(self.config.image_path), exist_ok=True)
-            logger.info(f"creating {disk_size_gb}GB storage for VM {self.config.name} at {self.config.image_path}")
-            self.hostconn.run_or_die(f'qemu-img create -f qcow2 {options} {self.config.image_path} {disk_size_gb}G')
-
-            cdrom_line = f"--cdrom {iso_or_image_path}"
-            append = "--wait=-1"
-            self.install_wait = True
-        else:
-            cdrom_line = ""
-            append = "--noautoconsole"
-            self.install_wait = False
-
-        if self.hostconn.is_localhost():
-            network = "network=default"
-        else:
-            network = "bridge=virbr0"
-        cmd = f"""
-        virt-install
-            --connect qemu:///system
-            -n {self.config.name}
-            -r {self.config.ram}
-            --cpu host
-            --vcpus {self.config.cpu}
-            --os-variant={self.config.os_variant}
-            --import
-            --network {network},mac={self.config.mac}
-            --events on_reboot=restart
-            {cdrom_line}
-            --disk path={self.config.image_path}
-            {append}
-        """
-
-        logger.info(f"Starting VM {self.config.name}")
-        ret = self.hostconn.run(cmd)
-        if ret.returncode != 0:
-            logger.info(f"Finished starting VM {self.config.name}, cmd = {cmd}, ret = {ret}")
-        else:
-            logger.info(f"Finished starting VM {self.config.name} without any errors")
-        return ret
-
-    def start(self, iso_or_image_path: str, executor: ThreadPoolExecutor) -> None:
-        self.future = executor.submit(self.setup_vm, iso_or_image_path)
-
-    def has_booted(self) -> bool:
-        return self.hostconn.vm_is_running(self.config.name)
-
-    def post_boot(self, desired_ip_range: Tuple[str, str]) -> bool:
-        if not self.install_wait:
-            self.future.result()
-        return True
-
-    def teardown(self) -> None:
-        # remove the image only if it really exists
-        image_path = self.config.image_path
-        self.hostconn.remove(image_path.replace(".qcow2", ".img"))
-        self.hostconn.remove(image_path)
-
-        # destroy the VM only if it really exists
-        if self.hostconn.run(f"virsh desc {self.config.name}").returncode == 0:
-            r = self.hostconn.run(f"virsh destroy {self.config.name}")
-            logger.info(r.err if r.err else r.out.strip())
-            r = self.hostconn.run(f"virsh undefine {self.config.name}")
-            logger.info(r.err if r.err else r.out.strip())
-
-
-class X86ClusterNode(ClusterNode):
-    external_port: str
-
-    def __init__(self, config: NodeConfig, external_port: str):
-        super().__init__(config)
-        self.external_port = external_port
-
-    def _boot_iso_x86(self, iso: str) -> host.Result:
-        logger.info(f"trying to boot {self.config.node} using {iso}")
-
-        lh = host.LocalHost()
-        nfs = NFS(lh, self.external_port)
-
-        bmc = BMC.from_bmc(self.config.bmc, self.config.bmc_user, self.config.bmc_password)
-        h = host.HostWithBF2(self.config.node, bmc)
-
-        iso = nfs.host_file(os.path.join(os.getcwd(), iso))
-        h.boot_iso_redfish(iso)
-        h.ssh_connect("core")
-        logger.info("connected")
-        return h.run("hostname")
-
-    def start(self, iso_or_image_path: str, executor: ThreadPoolExecutor) -> None:
-        self.future = executor.submit(self._boot_iso_x86, iso_or_image_path)
-
-    def has_booted(self) -> bool:
-        return self.future.done()
-
-    def post_boot(self, desired_ip_range: Tuple[str, str]) -> bool:
-        rh = host.RemoteHost(self.config.node)
-        rh.ssh_connect("core")
-        ipr_entries = common.ipa_to_entries(rh.run("ip -json a").out)
-        ips = []
-        for ipr in ipr_entries:
-            for addr_info in ipr.addr_info:
-                if addr_info.family != "inet":
-                    continue
-                if common.ip_range_contains(desired_ip_range, addr_info.local):
-                    ips.append(addr_info.local)
-
-        if len(ips) != 1:
-            logger.debug(f"Node {self.config.name} has unexpected IP addresses in range {desired_ip_range}.  Got: {ips}")
-            return False
-        self.dynamic_ip = ips[0]
-        return True
-
-
-class BFClusterNode(ClusterNode):
-    external_port: str
-
-    def __init__(self, config: NodeConfig, external_port: str):
-        super().__init__(config)
-        self.external_port = external_port
-
-    def _boot_iso_bf(self, iso: str) -> host.Result:
-        lh = host.LocalHost()
-        nfs = NFS(lh, self.external_port)
-
-        logger.info(f"Preparing BF on host {self.config.node}")
-        bmc = BMC.from_bmc(self.config.bmc, self.config.bmc_user, self.config.bmc_password)
-        h = host.HostWithBF2(self.config.node, bmc)
-        skip_boot = False
-        if h.ping():
-            try:
-                h.ssh_connect("core")
-                skip_boot = h.running_fcos()
-            except paramiko.ssh_exception.AuthenticationException:
-                logger.info("Authentication failed, will not be able to skip boot")
-
-        if skip_boot:
-            logger.info(f"Skipping booting {self.config.node}, already booted with FCOS")
-        else:
-            nfs_file = nfs.host_file("/root/iso/fedora-coreos.iso")
-            h.boot_iso_redfish(nfs_file)
-            time.sleep(10)
-            h.ssh_connect("core")
-
-        if not h.running_fcos():
-            logger.error_and_exit("Expected FCOS after booting host {self.config.node} but booted something else")
-
-        nfs_iso = nfs.host_file(f"/root/iso/{iso}")
-        nfs_key = nfs.host_file("/root/iso/ssh_priv_key")
-        output = h.bf_pxeboot(nfs_iso, nfs_key)
-        logger.debug(output)
-        if output.returncode:
-            logger.error_and_exit(f"Failed to run pxeboot on bf {self.config.node}")
-        else:
-            logger.info(f"succesfully ran pxeboot on bf {self.config.node}")
-
-        # ip is printed as the last thing when bf is pxeboot'ed
-        bf_ip = output.out.strip().split("\n")[-1].strip()
-        h.connect_to_bf(bf_ip)
-        max_tries = 3
-        bf_interfaces = ["enp3s0f0", "enp3s0f0np0"]
-        logger.info(f'Will try {max_tries} times to get an IP on {" or ".join(bf_interfaces)}')
-        ip = None
-        tries = 0
-        while True:
-            ipa = h.run_on_bf("ip -json a").out
-            detected = common.ipa_to_entries(ipa)
-            found = [e for e in detected if e.ifname in bf_interfaces]
-            if len(found) != 1:
-                logger.error(f"Failed to find expected number of interfaces on bf {self.config.node}")
-                logger.error(f"Output was: {ipa}")
-                sys.exit(-1)
-
-            ip = None
-            for e in found[0].addr_info:
-                if e.family == "inet":
-                    ip = e.local
-            if ip is not None:
-                break
-            logger.info(f"IP missing on {found[0]}, output was {ipa}")
-            tries += 1
-            if tries >= max_tries:
-                logger.error(f"IP missing on {found[0]}")
-                break
-            time.sleep(10)
-
-        if ip is None:
-            return host.Result(out="", err="Could not detect IP", returncode=1)
-        logger.info(f"Detected ip {ip}")
-        return host.Result(out=f"{ip}", err="", returncode=0)
-
-    def start(self, iso_or_image_path: str, executor: ThreadPoolExecutor) -> None:
-        self.future = executor.submit(self._boot_iso_bf, iso_or_image_path)
-
-    def has_booted(self) -> bool:
-        return self.future.done()
-
-    def post_boot(self, desired_ip_range: Tuple[str, str]) -> bool:
-        result: Optional[host.Result] = self.future.result()
-        if result is not None:
-            self.dynamic_ip = result.out
-        else:
-            logger.error_and_exit(f"Couldn't find ip of worker {self.config.name}")
-        return True
-
-
-class ClusterHost:
-    """
-    Physical host representation.  Contains fields and methods that allow to:
-    - connect to the host
-    - provision and tear down master/worker k8s nodes on the host
-    - configure host networking
-    """
-
-    hostconn: host.Host
-    bridge: VirBridge
-    config: HostConfig
-    k8s_master_nodes: List[ClusterNode]
-    k8s_worker_nodes: List[ClusterNode]
-    hosts_vms: bool
-
-    def __init__(self, h: host.Host, c: HostConfig, cc: ClustersConfig):
-        self.bridge = VirBridge(h)
-        self.hostconn = h
-        self.config = c
-
-        def _create_k8s_nodes(configs: List[NodeConfig]) -> List[ClusterNode]:
-            nodes: List[ClusterNode] = []
-            for node_config in configs:
-                if node_config.node != self.config.name:
-                    continue
-                if node_config.kind == "vm":
-                    nodes.append(VmClusterNode(self.hostconn, node_config))
-                elif node_config.kind == "physical":
-                    nodes.append(X86ClusterNode(node_config, cc.external_port))
-                elif node_config.kind == "bf":
-                    nodes.append(BFClusterNode(node_config, cc.external_port))
-                else:
-                    raise ValueError()
-            return nodes
-
-        self.k8s_master_nodes = _create_k8s_nodes(cc.masters)
-        self.k8s_worker_nodes = _create_k8s_nodes(cc.workers)
-        self.hosts_vms = any(k8s_node.config.kind == "vm" for k8s_node in self._k8s_nodes())
-
-        if not self.config.pre_installed:
-            self.hostconn.need_sudo()
-
-        if self.hosts_vms and not self.hostconn.is_localhost():
-            self.hostconn.ssh_connect(self.config.username, self.config.password)
-
-    def _k8s_nodes(self) -> List[ClusterNode]:
-        return self.k8s_master_nodes + self.k8s_worker_nodes
-
-    def _ensure_images(self, local_iso_path: str, infra_env: str, nodes: List[ClusterNode]) -> None:
-        if not self.hosts_vms:
-            return
-
-        image_paths = (node.config.image_path for node in nodes)
-        for vm_image_path in image_paths:
-            image_path = os.path.dirname(vm_image_path)
-            self.hostconn.run(f"mkdir -p {image_path}")
-            self.hostconn.run(f"chmod a+rw {image_path}")
-            iso_path = os.path.join(image_path, f"{infra_env}.iso")
-            logger.info(f"Copying {local_iso_path} to {self.hostconn.hostname()}:/{iso_path}")
-            self.hostconn.copy_to(local_iso_path, iso_path)
-            logger.debug(f"iso_path is now {iso_path} for {self.hostconn.hostname()}")
-
-    def configure_bridge(self) -> None:
-        if not self.hosts_vms:
-            return
-
-        self.bridge.configure(self.config.network_api_port)
-
-    def ensure_linked_to_network(self) -> None:
-        if not self.hosts_vms:
-            return
-
-        api_network = self.config.network_api_port
-        logger.info(f"link {api_network} to virbr0")
-        interface = common.find_port(self.hostconn, api_network)
-        if not interface:
-            logger.error_and_exit(f"Host {self.config.name} misses API network interface {api_network}")
-
-        br_name = "virbr0"
-
-        if interface.master is None:
-            logger.info(f"No master set for interface {api_network}, setting it to {br_name}")
-            self.hostconn.run(f"ip link set {api_network} master {br_name}")
-        elif interface.master != br_name:
-            logger.error_and_exit(f"Incorrect master set for interface {api_network}")
-
-        logger.info(f"Setting interface {api_network} as unmanaged in NetworkManager")
-        self.hostconn.run(f"nmcli device set {api_network} managed no")
-
-    def ensure_not_linked_to_network(self) -> None:
-        if not self.hosts_vms:
-            return
-
-        intif = self.validate_api_port()
-        if not intif:
-            logger.info("can't find network API port")
-        else:
-            logger.info(self.hostconn.run(f"ip link set {intif} nomaster"))
-            logger.info(f"Setting interface {intif} as managed in NetworkManager")
-            self.hostconn.run(f"nmcli device set {intif} managed yes")
-
-    def validate_api_port(self) -> Optional[str]:
-        if self.config.network_api_port == "auto":
-            self.config.network_api_port = common.get_auto_port(self.hostconn)
-
-        port = self.config.network_api_port
-        logger.info(f'Validating API network port {port}')
-        if not self.hostconn.port_exists(port):
-            logger.error(f"Can't find API network port {port}")
-            return None
-        if not self.hostconn.port_has_carrier(port):
-            logger.error(f"API network port {port} doesn't have a carrier")
-            return None
-        return port
-
-    def _configure_dhcp_entries(self, dhcp_bridge: VirBridge, nodes: List[ClusterNode]) -> None:
-        if not self.hosts_vms:
-            return
-
-        for node in nodes:
-            dhcp_bridge.setup_dhcp_entry(node.config)
-
-    def configure_master_dhcp_entries(self, dhcp_bridge: VirBridge) -> None:
-        return self._configure_dhcp_entries(dhcp_bridge, self.k8s_master_nodes)
-
-    def configure_worker_dhcp_entries(self, dhcp_bridge: VirBridge) -> None:
-        return self._configure_dhcp_entries(dhcp_bridge, self.k8s_worker_nodes)
-
-    def preinstall(self, external_port: str, executor: ThreadPoolExecutor) -> Future[None]:
-        def _preinstall() -> None:
-            if self.config.is_preinstalled():
-                return
-
-            iso = "fedora-coreos.iso"
-            coreosBuilder.ensure_fcos_exists(os.path.join(os.getcwd(), iso))
-            logger.debug(f"Provisioning Host {self.config.name}")
-
-            # Use the X86 node provisioning infrastructure to provision the host
-            # too.
-            x86_node = X86ClusterNode(self.k8s_worker_nodes[0].config, external_port)
-            x86_node._boot_iso_x86(iso)
-
-        return executor.submit(_preinstall)
-
-    def _start_nodes(self, iso_path: str, infra_env: str, executor: ThreadPoolExecutor, nodes: List[ClusterNode]) -> List[Future[Optional[host.Result]]]:
-        self._ensure_images(iso_path, infra_env, nodes)
-        futures = []
-        for node in nodes:
-            remote_iso_path = os.path.join(os.path.dirname(node.config.image_path), f"{infra_env}.iso")
-            node.start(remote_iso_path, executor)
-            futures.append(node.future)
-        return futures
-
-    def start_masters(self, iso_path: str, infra_env: str, executor: ThreadPoolExecutor) -> List[Future[Optional[host.Result]]]:
-        return self._start_nodes(iso_path, infra_env, executor, self.k8s_master_nodes)
-
-    def start_workers(self, iso_path: str, infra_env: str, executor: ThreadPoolExecutor) -> List[Future[Optional[host.Result]]]:
-        return self._start_nodes(iso_path, infra_env, executor, self.k8s_worker_nodes)
-
-    def _wait_for_boot(self, nodes: List[ClusterNode], desired_ip_range: Tuple[str, str]) -> None:
-        if not nodes:
-            return
-
-        try_count = 0
-        while True:
-            try_count += 1
-            states = {node.config.name: node.has_booted() for node in nodes}
-            node_names = ', '.join(states.keys())
-            logger.info(f"Waiting for nodes ({node_names}) to have booted (try #{try_count})..")
-            logger.info(f"Current boot state: {states}")
-            if all(has_booted for has_booted in states.values()):
-                break
-            time.sleep(10)
-        logger.info(f"It took {try_count} tries to wait for nodes ({node_names}) to have booted.")
-
-        try_count = 0
-        while True:
-            try_count += 1
-            states = {node.config.name: node.post_boot(desired_ip_range) for node in nodes}
-            node_names = ', '.join(states.keys())
-            logger.info(f"Waiting for nodes ({node_names}) to have run post_boot (try #{try_count})..")
-            logger.info(f"Current post_boot state: {states}")
-            if all(has_post_booted for has_post_booted in states.values()):
-                break
-            time.sleep(10)
-        logger.info(f"It took {try_count} tries to wait for nodes ({node_names}) to have run post_boot.")
-
-        for node in nodes:
-            node.health_check()
-
-    def wait_for_masters_boot(self, desired_ip_range: Tuple[str, str]) -> None:
-        return self._wait_for_boot(self.k8s_master_nodes, desired_ip_range)
-
-    def wait_for_workers_boot(self, desired_ip_range: Tuple[str, str]) -> None:
-        return self._wait_for_boot(self.k8s_worker_nodes, desired_ip_range)
-
-    def teardown(self) -> None:
-        for node in self._k8s_nodes():
-            node.teardown()
+from clusterHost import ClusterHost
 
 
 def match_to_proper_version_format(version_cluster_config: str) -> str:
